+++ frontend/src/actions/download.js
@@ -1,0 +1,6 @@
+export const CHART_DOWNLOAD_STATUS_UPDATE = 'CHART_DOWNLOAD_STATUS_UPDATE';
+export const updateChartDownloadStatus = (projectId, chartDownloadStatus) => ({
+  type: CHART_DOWNLOAD_STATUS_UPDATE,
+  projectId,
+  chartDownloadStatus,
+});
